diarization/requirements.txt
pyannote.audio==3.3.1
intervaltree==3.1.0
